# docs/swagger.yaml
basePath: /
definitions:
  api.AnagramRequest:
    properties:
      type:
        description: latin, runeglish, rune
        type: string
      word:
        type: string
    type: object
  api.AnagramResponse:
    properties:
      anagrams:
        items:
          type: string
        type: array
      count:
        type: integer
    type: object
  api.Base64ToCSVRequest:
    properties:
      base64:
        type: string
    type: object
  api.BytesToCSVResponse:
    properties:
      csv:
        type: string
    type: object
  api.CSVToBase64Request:
    properties:
      csv:
        type: string
    type: object
  api.CSVToBase64Response:
    properties:
      base64:
        type: string
    type: object
  api.CSVToBytesRequest:
    properties:
      csv:
        type: string
    type: object
...
  api.DictionaryWordsByParamRequest:
    properties:
      filters:
        items:
          properties:
            field:
              type: string
            param:
              type: integer
          type: object
        type: array
    type: object
  api.DictionaryWordsByParamResponse:
    properties:
      count:
        type: integer
      words:
        items:
          $ref: '#/definitions/tables.DictionaryWord'
        type: array
    type: object
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  api.GraphRequest:
    properties:
      points:
        items:
          $ref: '#/definitions/api.Point'
        type: array
    type: object
  api.GraphResponse:
    properties:
      lines:
        items:
          $ref: '#/definitions/api.Line'
        type: array
      points:
        items:
          $ref: '#/definitions/api.Point'
        type: array
      totalLength:
        type: number
    type: object
...
  api.Line:
    properties:
      length:
        type: number
      p1:
        $ref: '#/definitions/api.Point'
      p2:
        $ref: '#/definitions/api.Point'
    type: object
...
  api.Point:
    properties:
      x:
        type: number
      "y":
        type: number
    type: object
...
  tables.DictionaryWord:
    properties:
      dictRuneNoDoubletLength:
        type: integer
      dictionaryWordLength:
        type: integer
      dictionaryWordText:
        type: string
      gemProduct:
        type: string
      gemProductPrime:
        type: boolean
      gemSum:
        type: integer
      gemSumPrime:
        type: boolean
      language:
        type: string
      runeDistancePattern:
        type: string
      runePattern:
        type: string
      runePatternNoDoubletPattern:
        type: string
      runeWordLength:
        type: integer
      runeWordText:
        type: string
      runeWordTextNoDoublet:
        type: string
      runeglishWordLength:
        type: integer
      runeglishWordText:
        type: string
    type: object
host: localhost:3301
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Quantix Math API documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Quantix Math API
  version: "1.0"
paths:
  /api/binutils/base64-to-csv:
    post:
      consumes:
      - application/json
      description: Converts a Base64 string to a hex-encoded CSV string
      parameters:
      - description: Base64 to CSV Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.Base64ToCSVRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BytesToCSVResponse'
        "400":
          description: Invalid request body or Base64 format
          schema:
            type: string
      summary: Convert Base64 to hex CSV
      tags:
      - Utilities
  /api/binutils/csv-to-base64:
    post:
      consumes:
      - application/json
      description: Converts a hex-encoded CSV string to a Base64 string
      parameters:
      - description: CSV to Base64 Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CSVToBase64Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CSVToBase64Response'
        "400":
          description: Invalid request body or CSV format
          schema:
            type: string
      summary: Convert hex CSV to Base64
      tags:
      - Utilities
  /api/binutils/csv-to-bytes:
    post:
      consumes:
      - application/json
      description: Converts a hex-encoded CSV string back to binary data and returns
        it as a file
      parameters:
      - description: CSV to Bytes Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CSVToBytesRequest'
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Invalid request body or CSV format
          schema:
            type: string
      summary: Convert hex CSV to binary
      tags:
      - Utilities
  /api/binutils/file-to-csv:
    post:
      consumes:
      - multipart/form-data
      description: Reads a file and returns its content as a hex-encoded CSV string
      parameters:
      - description: File to convert
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BytesToCSVResponse'
        "400":
          description: File upload required
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Convert a file to hex CSV
      tags:
      - Utilities
...
  /api/dictionary/anagrams:
    post:
      consumes:
      - application/json
      description: Returns a list of anagrams for the given word and type
      parameters:
      - description: Anagram Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AnagramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AnagramResponse'
        "400":
          description: Invalid request body or word
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Find anagrams for a word
      tags:
      - Dictionary
  /api/dictionary/words:
    post:
      consumes:
      - application/json
      description: Returns a list of dictionary words based on the provided filters
      parameters:
      - description: Search Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.DictionaryWordsByParamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DictionaryWordsByParamResponse'
        "400":
          description: Invalid request body or filters
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Search for dictionary words
      tags:
      - Dictionary
  /api/dictionary/words/download:
    get:
      description: Generates and returns an Excel file containing dictionary words
        based on filters
      parameters:
      - collectionFormat: csv
        description: Filter fields (e.g., dict_word_length, gem_sum)
        in: query
        items:
          type: string
        name: field
        required: true
        type: array
      - collectionFormat: csv
        description: Filter values
        in: query
        items:
          type: integer
        name: param
        required: true
        type: array
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Invalid or mismatched parameters
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Download dictionary words as Excel
      tags:
      - Dictionary
  /api/graph/process:
    post:
      consumes:
      - application/json
      description: Calculates lines between points and total length, filtering lines
        that pass through other points
      parameters:
      - description: Graph Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.GraphRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GraphResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Process a set of coordinate points
      tags:
      - Math
...
swagger: "2.0"